+++ data.php
@@ -1,0 +1,22 @@
+<?php 
+$is_auth = (bool) rand(0, 1);
+
+//$user_name = 'Константин';
+$user_avatar = 'img/user.jpg';
+
+// устанавливаем часовой пояс в Московское время
+date_default_timezone_set('Europe/Moscow');
+
+// записать в эту переменную оставшееся время в этом формате (ЧЧ:ММ)
+$lot_time_remaining = "00:00";
+
+// временная метка для полночи следующего дня
+$tomorrow = strtotime('tomorrow midnight');
+
+// временная метка для настоящего времени
+$now = strtotime('now');
+
+$categories = ['Доски и лыжи','Крепления','Ботинки','Одежда','Инструменты','Разное'];
+
+$lot_time_remaining = time_add_zero($now, $tomorrow);
+?>
