+++ logout.php
@@ -1,0 +1,9 @@
+<?php
+session_start();
+
+unset($_SESSION['user']);
+header("Location: /index.php");
+
+
+// разлогинивание
+?>
