+++ lots_list.php
@@ -43,4 +43,11 @@ $equipment = [
 'description' => 'Описание отсутствует'
 ]
 ];
+
+$bets = [
+	['name' => 'Иван', 'price' => 11500, 'ts' => strtotime('-' . rand(1, 50) .' minute')],
+	['name' => 'Константин', 'price' => 11000, 'ts' => strtotime('-' . rand(1, 18) .' hour')],
+	['name' => 'Евгений', 'price' => 10500, 'ts' => strtotime('-' . rand(25, 50) .' hour')],
+	['name' => 'Семён', 'price' => 10000, 'ts' => strtotime('last week')]
+];
 ?>
