+++ userdata.php
@@ -18,3 +18,4 @@ $users = [
         'password' => '$2y$10$2OxpEH7narYpkOT1H5cApezuzh10tZEEQ2axgFOaKW.55LxIJBgWW'
     ]
 ];
+
